D2: physically move pure quest logic into quest/domain and drop reactflow
Replace the quest/domain re-export shims with the real implementations:
the graph pipeline (node identification, edge building, layout, shared
helpers/types/constants), command executors, guardrails, and analysis now
live under src/renderer/quest/domain/.

Define editor-owned QuestGraphNode/QuestGraphEdge interfaces in
types/questGraph.ts instead of reactflow's Node/Edge, and replace
MarkerType with a plain literal in questEdgeBuilding. Nothing imports
reactflow anymore, so the dependency, its vite chunk rule, and the
vestigial jest mock are removed.

The layering contract (UI -> application -> domain, no UI libraries in
the domain) is now enforced by tests/questDomainBoundary.test.ts.
