--- daedalus-dialog-editor/src/renderer/quest/domain/commands.ts
@@ -1,2 +1,0 @@
-export { executeQuestGraphCommand } from '../../components/QuestEditor/commands';
-export type { QuestGraphCommand } from '../../components/QuestEditor/commands';
