+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/addLogEntry.ts
