+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/connectCondition.ts
