+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/index.ts
