+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/removeTransition.ts
