+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/shared.ts
