+++ daedalus-dialog-editor/src/renderer/quest/domain/commands/types.ts
