+++ daedalus-dialog-editor/src/renderer/quest/domain/questGraphConstants.ts
