+++ daedalus-dialog-editor/src/renderer/quest/domain/questGraphInternalTypes.ts
