+++ daedalus-dialog-editor/src/renderer/quest/domain/questGraphSharedHelpers.ts
