+++ daedalus-dialog-editor/src/renderer/quest/domain/questNodeIdentification.ts
