+++ docs/refactoring-targets.md
@@ -21,22 +21,7 @@ Audit of god components, muddled concerns, and maintainability red flags.
 
 ---
 
-### 3. Quest domain layer — physically move logic out of the UI tree
-**Files:** `daedalus-dialog-editor/src/renderer/quest/domain/*`,
-`.../components/QuestEditor/*`, `.../types/questGraph.ts`
-
-The `quest/domain/*` modules are currently re-export shims pointing back into
-`components/QuestEditor/*` (see `docs/architecture/quest-editor.md`). To make the
-documented one-way layering physically true:
-
-- Move the pure pipeline/command/guardrail/analysis logic into `quest/domain/*`.
-- Remove the `reactflow` dependency from the domain by defining editor-owned
-  graph node/edge types (replacing `reactflow`'s `Node`/`Edge` in
-  `types/questGraph.ts` and `MarkerType` in `questEdgeBuilding.ts`).
-
----
-
-### 4. History snapshots — avoid full deep clones
+### 3. History snapshots — avoid full deep clones
 **File:** `daedalus-dialog-editor/src/renderer/store/historyStore.ts`,
 `.../utils/historyUtils.ts`
 
@@ -47,7 +32,7 @@ test pass for aliasing safety.
 
 ---
 
-### 5. `properties.information`/`condition` string-vs-object union
+### 4. `properties.information`/`condition` string-vs-object union
 **Files:** parser model types + `daedalus-dialog-editor/src/renderer/**`
 
 The dialog property references are sometimes a plain string and sometimes a
